+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - server: explaining/analyzing a query now works for mssql sources
 - server: fix MSSQL multiplexed subscriptions (fix #6887)
 - server: fix bug preventing tables with the same name in different sources
+- server: include more detail in inconsistent metadata error messages (fix #6684)
 - console: add union types to remote schema permissions
 - console: read-only modify page for mssql
 - console: filter out partitions from track table list and display partition info